feat(http)!: widen maxRequestBodySize to long
The maxRequestBodySize map and HTTPInputStream's content-length tracking
were int-typed, capping configurable body sizes at 2 GiB. MultipartConfiguration's
maxFileSize was already long, so the cross-check in HTTPServer.validateConfiguration
silently narrowed at the int boundary. Widen the API and internals end-to-end.

+++ src/main/java/org/lattejava/http/server/Configurable.java
@@ -224,7 +224,7 @@ public interface Configurable<T extends Configurable<T>> {
    * @param maxRequestBodySize a map specifying the maximum size in bytes for the HTTP request body by Content-Type
    * @return This.
    */
-  default T withMaxRequestBodySize(Map<String, Integer> maxRequestBodySize) {
+  default T withMaxRequestBodySize(Map<String, Long> maxRequestBodySize) {
     configuration().withMaxRequestBodySize(maxRequestBodySize);
     return (T) this;
   }
